Fix issues #128 #129 + improvements

+++ scripts/install_vsftpd.sh
@@ -163,16 +163,20 @@ function init_vsftpd_install() {
         echo "Configuring FTP server (VSFTPD)..."
 
         if [[ "${DRYRUN}" != true ]]; then
+            FTP_MIN_PORT=${FTP_MIN_PORT:-45000}
+            FTP_MAX_PORT=${FTP_MAX_PORT:-45099}
+
             # Backup default vsftpd conf.
-            [[ -f /etc/vsftpd.conf ]] && \
+            if [[ -f /etc/vsftpd.conf ]]; then
                 run mv /etc/vsftpd.conf /etc/vsftpd.conf.bak
-    
+            fi
+
             run touch /etc/vsftpd.conf
 
-            # Enable jail
+            # Enable jail mode.
             cat > /etc/vsftpd.conf <<EOL
-listen=NO
-listen_ipv6=YES
+listen=YES
+listen_ipv6=NO
 anonymous_enable=NO
 local_enable=YES
 write_enable=YES
@@ -188,14 +192,21 @@ pam_service_name=vsftpd
 force_dot_files=YES
 
 pasv_enable=YES
-pasv_min_port=40000
-pasv_max_port=50000
+pasv_min_port=${FTP_MIN_PORT}
+pasv_max_port=${FTP_MAX_PORT}
+#pasv_address=${SERVER_IP}
+#pasv_addr_resolve=YES
 
-user_sub_token=${USER}
-local_root=/home/${USER}
+user_sub_token=\$USER
+local_root=/home/\$USER
+
+userlist_enable=YES
+userlist_file=/etc/vsftpd.userlist
+userlist_deny=NO
 EOL
 
             # Enable SSL.
+            # TODO: Change the self-signed certificate with a valid Let's Encrypt certificate.
             if [[ "${VSFTPD_SSL_ENABLE}" == true ]]; then
                 cat >> /etc/vsftpd.conf <<EOL
 
@@ -213,6 +224,23 @@ rsa_cert_file=/etc/ssl/certs/ssl-cert-snakeoil.pem
 rsa_private_key_file=/etc/ssl/private/ssl-cert-snakeoil.key
 EOL
             fi
+
+            # If using elastic IP (such as AWS EC2), set the server IP.
+            if [[ "${SERVER_IP}" != "$(get_ip_private)" ]]; then
+                run sed -i "s|^#pasv_address=.*|pasv_address=${SERVER_IP}|g" /etc/vsftpd.conf
+                run sed -i "s|^#pasv_addr_resolve=.*|pasv_addr_resolve=YES|g" /etc/vsftpd.conf
+            fi
+
+            # If Let's Encrypt SSL certificate is issued for hostname, set the certificate.
+            if [[ -n "${HOSTNAME_CERT_PATH}" && -f "${HOSTNAME_CERT_PATH}/fullchain.pem" ]]; then
+                run sed -i "s|^rsa_cert_file=[^[:digit:]]*$|rsa_cert_file=${HOSTNAME_CERT_PATH}/fullchain.pem|g" /etc/vsftpd.conf
+                run sed -i "s|^rsa_private_key_file=[^[:digit:]]*$|rsa_private_key_file=${HOSTNAME_CERT_PATH}/privkey.pem|g" /etc/vsftpd.conf
+            fi
+
+            # Add default LEMPer Stack user to vsftpd.userlist.
+            LEMPER_USERNAME=${LEMPER_USERNAME:-lemper}
+            run touch /etc/vsftpd.userlist
+            run bash -c "echo '${LEMPER_USERNAME}' | tee -a /etc/vsftpd.userlist"
         fi
 
         # Add systemd service.
@@ -221,7 +249,7 @@ EOL
         [[ ! -f /etc/systemd/system/multi-user.target.wants/vsftpd.service ]] && \
             run ln -s /lib/systemd/system/vsftpd.service /etc/systemd/system/multi-user.target.wants/vsftpd.service
 
-        # Restart Fail2ban daemon.
+        # Restart vsftpd daemon.
         echo "Restarting FTP server (VSFTPD)..."
         run systemctl unmask vsftpd
         run systemctl restart vsftpd
